refactor: extract computed and watchers
